#101: Unify Input and Output types into a common Span type, but keeping Input and Output as strong separated types
