+++ include/k3/tok3n/types.h
@@ -1,10 +1,9 @@
 #pragma once
 
-#include <k3/tok3n/types/Input.h>
 #include <k3/tok3n/types/IntegralConstants.h>
 #include <k3/tok3n/types/ModifierBase.h>
-#include <k3/tok3n/types/Output.h>
 #include <k3/tok3n/types/ParserType.h>
 #include <k3/tok3n/types/Result.h>
+#include <k3/tok3n/types/Span.h>
 #include <k3/tok3n/types/StaticString.h>
 #include <k3/tok3n/types/Tags.h>
